update env.example and move version handler
