+++ deployments/local-docker/.env.example
@@ -1,5 +1,11 @@
 AWS_LAMBDA_RUNTIME_API=
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_DEFAULT_REGION=
+AWS_ACCESSKEYID=
+AWS_SECRETACCESSKEY=
+AWS_DEFAULTREGION=
 AWS_BUCKET=
+EMAIL_SMTP_HOST=
+EMAIL_SMTP_PORT=
+EMAIL_SMTP_SECURE=
+EMAIL_SMTP_USERNAME=
+EMAIL_SMTP_PASSWORD=
+EMAIL_IAM_USERNAME=
