+++ internal/handlers/transaction.go
@@ -11,22 +11,6 @@ import (
 )
 
 /*
- * HandleVersion
- * Simple version handler to check if the API is running
- * @param w http.ResponseWriter
- * @param r *http.Request
- * @return void
- *
- */
-func HandleVersion(w http.ResponseWriter, r *http.Request) {
-	version := models.Version{}
-	version.Version = "1.0.1 - 2021-09-30 - transactions-history"
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(version)
-}
-
-/*
  * HandleTransactions
  * @param w http.ResponseWriter
  * @param r *http.Request
